requirements.txt
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
drf-yasg==1.21.7
drf-nested-routers==0.93.4